restructured lib/, added app/ folder
