+++ src/app/App.js
@@ -1,28 +1,33 @@
-import React, {useEffect, useRef} from 'react';
-import ReactGA from 'react-ga';
-import {createClickTracking} from './lib/helpers';
-import {BrowserRouter as Router, Route, Switch, Link} from 'react-router-dom';
-import {MuiThemeProvider} from '@material-ui/core/styles';
-import AppBar from '@material-ui/core/AppBar';
-import Button from '@material-ui/core/Button';
-import MenuItem from '@material-ui/core/MenuItem';
-import Toolbar from '@material-ui/core/Toolbar';
-import withStyles from '@material-ui/core/styles/withStyles';
+import React, {useEffect, useRef} from '../../node_modules/react';
+import ReactGA from '../../node_modules/react-ga';
+import {createClickTracking} from '../lib/helperFunctions/helpers';
+import {
+  BrowserRouter as Router,
+  Route,
+  Switch,
+  Link,
+} from '../../node_modules/react-router-dom';
+import {MuiThemeProvider} from '../../node_modules/@material-ui/core/styles';
+import AppBar from '../../node_modules/@material-ui/core/AppBar';
+import Button from '../../node_modules/@material-ui/core/Button';
+import MenuItem from '../../node_modules/@material-ui/core/MenuItem';
+import Toolbar from '../../node_modules/@material-ui/core/Toolbar';
+import withStyles from '../../node_modules/@material-ui/core/styles/withStyles';
 
-import HomeIcon from '@material-ui/icons/Home';
+import HomeIcon from '../../node_modules/@material-ui/icons/Home';
 
-import ErrorBoundary from 'atoms/ErrorBoundary';
-import Resume from 'components/Resume/Resume';
+import ErrorBoundary from '../lib/atoms/ErrorBoundary';
+import Resume from '../components/Resume/Resume';
 
-import {ProfileAuth, ProfileStaff} from 'components/ProfilePage';
-import theme from 'styles/theme';
+import {ProfileAuth, ProfileStaff} from '../components/ProfilePage';
+import theme from '../styles/theme';
 
-import {useAuth0} from 'lib/auth0';
+import {useAuth0} from '../lib/Auth0/auth0';
 
-import Home from 'components/Home/Home.js';
-import Contacts from 'components/Contacts/Contacts.container';
+import Home from '../components/Home/Home';
+import Contacts from '../components/Contacts/Contacts.container';
 
-import NavBarIcons from './components/NavigationBar/NavBarIcons';
+import NavBarIcons from '../components/NavigationBar/NavBarIcons';
 
 const App = ({
   hasSession,
