+++ src/components/Home/Home.js
@@ -10,9 +10,9 @@ import CardContent from '@material-ui/core/CardContent';
 import CardMedia from '@material-ui/core/CardMedia';
 import Button from '@material-ui/core/Button';
 import Typography from '@material-ui/core/Typography';
-import {createClickTracking} from '../../lib/helpers';
+import {createClickTracking} from '../../lib/helperFunctions/helpers';
 
-import {useAuth0} from 'lib/auth0';
+import {useAuth0} from 'lib/Auth0/auth0';
 
 const Home = ({hasSession, classes}) => {
   const {isAuthenticated, loginWithRedirect} = useAuth0();
