+++ src/components/ProfilePage/ProfilePage.js
@@ -21,7 +21,7 @@ import ExperiencesList from 'components/Experiences/ExperiencesList';
 import SkillsSection from 'components/Skill/SkillsSection';
 
 import HelpDrawer from '../SideBarDrawer/HelpDrawer';
-import {createExternalLink} from '../../lib/helpers';
+import {createExternalLink} from '../../lib/helperFunctions/helpers';
 
 import CAPABILITIES from './capabilities.yml';
