+++ src/components/Skill/SkillSelect/SkillSelect.container.js
@@ -1,7 +1,7 @@
 import 'isomorphic-fetch';
 import React from 'react';
 
-import {API_URL} from '../../../constants';
+import {API_URL} from '../../../app/constants';
 import throttle from 'lodash/throttle';
 
 import SkillSelect from './SkillSelect';
