+++ src/lib/Auth0/auth0.js
